+++ include/linux/swap.h
@@ -68,7 +68,7 @@ extern int nr_free_pages;
 extern atomic_t nr_async_pages;
 extern struct inode swapper_inode;
 extern unsigned long page_cache_size;
-extern int buffermem;
+extern long buffermem;
 
 /* Incomplete types for prototype declarations: */
 struct task_struct;
